requirements.txt
django>=6.0.2
djangorestframework
psycopg2-binary
requests
